multiemoVA database
I got the aus-es and trained the model but the accuracy is really low. Tryed to mix the two databases and train it with both but there is no good result because this one is really low. Trying to fix that. The model right know is trained with just diffusionFER. To change it just change the aus file in training_file