MultiEmoVA database
I pushed this before but idk what happened. Training with just diffusionFER for now because multiEmo has low accuracy. Created a way of merging the two data . Uploading that but right know is just with difusion. To try just need to change the name of the file. the general one is aus.